+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.10- 2018-06-09
+* Actually pass options
+* Allow _itunes extensions to be set/overridden from the itunes options object.
+
 ## 1.0.3- 2018-06-09
 * Finished up even more iTunes tag mappings and improved documentation.  Lots of small changes.